「优化」添加下载工具类 带进度 「优化」保存文件使用okio 「优化」FastJsonResponsebodyConverter代码优化 「优化」FileCallback文件丢弃使用新方案具体看注释 「优化」文件FileE代码优化

+++ okhttputils/src/main/java/com/jelly/thor/okhttputils/utils/file/FileInProgress.kt
@@ -10,5 +10,5 @@ abstract class FileInProgress constructor(val totalSize: Long) {
      * @param progress 进度
      * @param total 总大小
      */
-    abstract fun inProgress(progress: Float, total: Long)
+    abstract fun inProgress(progress: Long, total: Long)
 }
